+++ packages/edge-bundler/node/manifest.test.ts
@@ -1,4 +1,4 @@
-import { test, expect } from 'vitest'
+import { describe, test, expect } from 'vitest'
 
 // @ts-expect-error current tsconfig.json doesn't allow this, but I don't want to change it
 import { version } from '../package.json' assert { type: 'json' }
@@ -237,11 +237,11 @@ test('excludedPath from ISC goes into function_config, TOML goes into routes', (
 
   const matcher = getRouteMatcher(manifest)
 
-  expect(matcher('/showcases/boho-style')).toBeDefined()
+  expect(matcher('/showcases/boho-matcher')).toBeDefined()
   expect(matcher('/checkout/address')).toBeDefined()
   expect(matcher('/checkout/terms-and-conditions')).toBeUndefined()
   expect(matcher('/checkout/scrooge-mc-duck-animation.css')).toBeUndefined()
-  expect(matcher('/showcases/boho-style/expensive-chair.jpg')).toBeUndefined()
+  expect(matcher('/showcases/boho-matcher/expensive-chair.jpg')).toBeUndefined()
 })
 
 test('URLPattern named groups are supported', () => {
@@ -571,4 +571,98 @@ test('Generates a manifest with rewrite config', () => {
   }
   expect(manifest.routes).toEqual(expectedRoutes)
   expect(manifest.function_config).toEqual(expectedFunctionConfig)
+})
+
+describe('Header matching', () => {
+  test('Throws a bundling error if the type is incorrect', () => {
+    const functions = [{ name: 'func-1', path: '/path/to/func-1.ts' }]
+
+    expect(() =>
+      generateManifest({
+        bundles: [],
+
+        // @ts-expect-error Incorrect type
+        declarations: [{ function: 'func-1', path: '/f1/*', header: 'foo' }],
+        functions,
+      }),
+    ).toThrowError(BundleError)
+
+    expect(() =>
+      generateManifest({
+        bundles: [],
+
+        declarations: [
+          {
+            function: 'func-1',
+            path: '/f1/*',
+            header: {
+              'x-correct': true,
+
+              // @ts-expect-error Incorrect type
+              'x-not-correct': {
+                problem: true,
+              },
+            },
+          },
+        ],
+        functions,
+      }),
+    ).toThrowError(BundleError)
+  })
+
+  test('Writes header matching rules to the manifest', () => {
+    const functions = [{ name: 'func-1', path: '/path/to/func-1.ts' }]
+    const declarations: Declaration[] = [
+      {
+        function: 'func-1',
+        path: '/f1/*',
+        header: {
+          'x-present': true,
+          'x-also-present': true,
+          'x-absent': false,
+          'x-match-prefix': '^prefix',
+          'x-match-exact': '^exact$',
+          'x-match-suffix': 'suffix$',
+        },
+      },
+    ]
+    const { manifest } = generateManifest({
+      bundles: [],
+      declarations,
+      functions,
+    })
+
+    const expectedRoutes = [
+      {
+        function: 'func-1',
+        pattern: '^/f1(?:/(.*))/?$',
+        excluded_patterns: [],
+        path: '/f1/*',
+        headers: {
+          'x-absent': {
+            matcher: 'missing',
+          },
+          'x-also-present': {
+            matcher: 'exists',
+          },
+          'x-match-exact': {
+            pattern: '^exact$',
+            matcher: 'regex',
+          },
+          'x-match-prefix': {
+            pattern: '^prefix',
+            matcher: 'regex',
+          },
+          'x-match-suffix': {
+            pattern: 'suffix$',
+            matcher: 'regex',
+          },
+          'x-present': {
+            matcher: 'exists',
+          },
+        },
+      },
+    ]
+    expect(manifest.routes).toEqual(expectedRoutes)
+  })
 })